added recording interface

+++ theremin/theremin/InstrumentViewController.swift
@@ -25,6 +25,7 @@ class InstrumentViewController: UIViewController {
     
     let note_positions = ["C" : 0, "C#" : 1, "Db" : 1, "D" : 2, "D#" : 3, "Eb" : 3, "E" : 4, "Fb" : 4, "E#" : 5, "F" : 5, "F#" : 6, "Gb" : 6, "G" : 7, "G#" : 8, "Ab" : 8, "A" : 9, "A#" : 10, "Bb" : 10, "B" : 11, "B#" : 11, "Cb" : 11]
 
+
     var key_names = ["Major", "Minor"]
     
     var key_note: String = "C"
@@ -35,7 +36,7 @@ class InstrumentViewController: UIViewController {
 
     //var leftmost_note = "C"
     let num_oct: Int = 4
-    let bottom_note: Int = 59
+    let bottom_note: CGFloat = 59.0
     
     var grid: GridViewController!
     var range_controller: RangeViewContainerController!
